+++ app/src/main/java/vn/fpoly/fpolybookcarclient/view/maps/ViewGoogleMap.java
@@ -4,7 +4,7 @@ import vn.fpoly.fpolybookcarclient.model.objectClass.Driver;
 
 public interface ViewGoogleMap {
    void drawPolyline();
-   void showDetailDistance(int distance, int time,int price);
+   void showDetailDistance(int distance, int time,double price);
 
    void loadDriverCarList();
    void loadDriverMotoList();
